update webpack to allow dotenv dep injection, move keys to .env, update in gmaps and injection
