+++ .env.example
@@ -1,5 +1,5 @@
-SECRET=AsRP76XrEe2jJ3Xi4Js0oAsRP76XrEe2jJ3Xi4Js0oAsRP76XrEe2jJ3Xi4Js0oAsRP76XrEe2jJ3Xi4Js0oAsRP76XrEe2jJ3Xi4Js0o
-VERSION=v4.2
+SECRET=0oAsRP76XrEe2jJ3Xi4Js0oJs0oAsRPAsR6XrEe2jJ3Xi4P7XrEe2jJ3Xi4Js0oAsRP76XrEe2jJ3Xi4Js76XrEe2jJ3Xi4Js0oAsRP76
+VERSION=v4.3
 
 # webpack
 WEB_SERVER_HOST="localhost"
@@ -9,5 +9,8 @@ WEBPACK_SERVER_PORT="7002"
 
 # rack
 PORT="7000"
-
 MODE="dev"
+
+# keys
+CDNURL="//cloudfront/full/path"
+GMAPSJSAPIKEY="get-one-from-google"
