+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "unimelb-design-system-assets",
   "description": "Asset compilation using Webpack",
-  "version": "4.2.0",
+  "version": "4.3.0",
   "repository": "marcom-unimelb/unimelb-design-system",
   "author": "University of Melbourne",
   "license": "UNLICENSED",
@@ -27,7 +27,7 @@
     "babel-loader": "~5.3.2",
     "cssesc": "~0.1.0",
     "css-loader": "~0.9.1",
-    "dotenv": "~1.1.0",
+    "dotenv": "~2.0.0",
     "express": "~4.12.1",
     "extract-text-webpack-plugin": "~0.3.8",
     "file-loader": "~0.8.1",
@@ -39,8 +39,8 @@
     "node-sass": "~3.3.3",
     "sass-loader": "~3.0.0",
     "style-loader": "~0.8.3",
-    "webpack": "^1.6.0",
+    "webpack": "^1.13.1",
     "webpack-dev-middleware": "~1.0.11",
-    "webpack-dev-server": "~1.8.2"
+    "webpack-dev-server": "~1.14.1"
   }
 }
